alph build of numbergoup
