+++ next.config.js
@@ -7,4 +7,11 @@ module.exports = {
     experimental: {
       appDir: true,
     },
+    webpack: (config, { isServer }) => {
+      // Replace 'fs' module with an empty module on client side
+      if (!isServer) {
+        config.resolve.fallback = { fs: false };
+      }
+      return config;
+    },
   };
