--- pages/alphpaca_farm.tsx
@@ -1,28 +1,0 @@
-// Page alphpaca_farm.tsx
-import React from 'react'
-import Head from 'next/head'
-import styles from '@/styles/Home.module.css'
-import Image from 'next/image';
-import { TypeAnimation } from 'react-type-animation';
-import { useState, useEffect } from 'react';
-import Link from 'next/link';
-
-// Alephium Imports
-import { AlephiumConnectButton } from '@alephium/web3-react'
-import { AlephiumWalletProvider } from '@alephium/web3-react';
-import { NodeProvider } from '@alephium/web3';
-
-// Components
-import { Navbar } from '@/components/NavBar';
-import dynamic from 'next/dynamic';
-
-const GameWindow = dynamic(() => import('../components/GameWindow'), { ssr: false });
-
-export default function ALPHpaca() {
-    return (
-        <div>
-            <Navbar />
-            <GameWindow />
-        </div>
-    );
-}
