+++ src/pages/_app.tsx
@@ -1,15 +1,13 @@
 import React from 'react'
-import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
 import { AlephiumWalletProvider } from '@alephium/web3-react'
-import { TokenFaucetConfig } from '@/services/utils'
+import { SubscribeConfig } from '@/utils/utils'
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <AlephiumWalletProvider
-      useTheme="midnight"
-      network={TokenFaucetConfig.network}
-      addressGroup={TokenFaucetConfig.groupIndex}
+      network={SubscribeConfig.network}
+      addressGroup={SubscribeConfig.groupIndex}
     >
       <Component {...pageProps} />
     </AlephiumWalletProvider>
